+++ lib/240325/data_source/todo.dart
@@ -1,0 +1,27 @@
+class Todo {
+  final int userId;
+  final int id;
+  final String title;
+  final bool completed;
+
+  Todo({
+    required this.userId,
+    required this.id,
+    required this.title,
+    required this.completed,
+  });
+
+  factory Todo.fromJson(Map<String, dynamic> json) {
+    return Todo(
+      userId: json['userId'],
+      id: json['id'],
+      title: json['title'],
+      completed: json['completed'],
+    );
+  }
+
+  @override
+  String toString() {
+    return 'Userid = $userId id = $id title = $title completed = $completed';
+  }
+}
